Add a basic COM Service to the Updater.
This is a part of the effort to move the Updater core functionality into
a singleton service.

The updater core, which is the process managing the installed
application set and containing the calls to update_client, needs to run
as a singleton COM service on Windows so that it can be reached by
multiple clients, such as on-demand, metainstallers, etc.

BUG=1048653

Change-Id: Ic2f7f26dea04e7d427f76853cf011c5965bf93fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2051612
Commit-Queue: S. Ganesh <ganesh@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#742817}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7115044c30d2bae37f229dbbb173a44b7fb36737