Avoid using high contrast colors for controls on non-Windows platforms.
Prior to this CL when a high contrast like setting was enabled, it meant
that the high contrast colors were going to be used for drawing the controls.

For non-Windows platform, the high contrast colors are wrong, because there is
no support yet to map the system colors to the high contrast colors.

This CL restricts the usage of high contrast colors just for Windows where
we have the correct colors.

Bug: 1022118
Change-Id: I20dbd4b364619b759b4843f4e974332a6e1ffa3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2065397
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#743285}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4548d56e253c50616452512310a5b9043f28316f