Center OOPIF interest rect around viewport bounds, not visible rect
When an OOPIF is bigger than its parent frame, we limit the area of the
iframe to be rasterized to the size of its parent frame, plus some extra
space to prevent scroll guttering. In this situation, The existing code
centers this area to be composited around the part of the iframe that is
visible in the top-level document after all intervening clips have been
applied. This CL does not change the size of the rect, but it centers
the rect around the parent frame.

The motivation for this CL is that a subsequent CL is going to change
RemoteFrameView::GetCompositingRect() so that it isn't calculated on the
fly (which is the root cause of the linked bug). Instead, the
compositing rect will be computed during each lifecycle update, and
GetCompositingRect() will return a cached value computed during the
last lifecycle update.

In the subsequent CL, the compositing rect needs to be updated prior
to the parent frame committing to the compositor thread, because the
parent frame will submit the child frame's compositing bounds, and it
needs to be up-to-date. But the current logic to compute the compositing
bounds relies on the result of FrameView::UpdateViewportIntersection,
which doesn't run until after compositor commit.

So the whole motivation for this change is to break the dependency
on RemoteFrameView::last_intersection_state_ when computing the
compositing bounds.

Although the interest rect will no longer be centered around the
iframe's viewport intersection, the "extra" area intended to prevent
scroll guttering will still always be at least 15% of the parent frame's
size in each direction, so this should not regress any visible behavior.

BUG=1033746

Change-Id: I1aac0dea2dddaa60be12de2d3635bb77cbd55c52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2045223
Commit-Queue: Stefan Zager <szager@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#743388}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a3b23699da40ef820addf39c0dc3ea5dea73f5e