Calculate SafeOpaqueBackgroundColor for layer instead of for PaintChunk
This is a preparation for fixing hit test of black backgrounds in
CompositeAfterPaint (crbug.com/937573). The new method finds the
background color from the first drawable display item in a layer after
we paint the layer.

In case that the first paint chunk is just for a blank background of
an ancestor like:
  <div id="parent" style="will-change: transform">
    <div id="child" style="position: relative;
         height: 100px; background: blue"></div>
  </div>
we'll use the child's background color as the
safe_opaque_background_color of the layer created for parent.

This also avoids unnecessary safe_opaque_background_color in
PaintChunks that are not used to set that of the layer.

Bug: 937573
Change-Id: Idf8e1c8fd9bb49cfd83fec41964589c28d9b4eba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068640
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#743701}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 02dd26fc4eef54b7428547bd339605900c2f9a00