WebApps: always bind WebLaunchService for File Handling
Prior to this CL, WebLaunchServiceImpl is behind a FileHandlingEnabled
check on the document. However, at the time of this check, the actual
document is not loaded (we get a blank document without origin trial
tokens). So FileHandlingEnabled check always return false, unless we
turn on runtime flags.

This check cause the problem: when a PWA with File Handling origin
trial is launched as a file handler, the app will not receive the
launch files.

This CL fixes the problem by always bind WebLaunchServiceImpl, this is
okay because the DOM (JavaScript) side is still behind a origin trial
check, we will not be exposing extra APIs to webpages.

Fixed: 1053819
Bug: 1045297
Change-Id: I19966d26afe97dfd3115f142343c6eaacdfe0eab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2063196
Commit-Queue: Jiewei Qian  <qjw@chromium.org>
Reviewed-by: Olivier Yiptong <oyiptong@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#743789}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 38926cb5843fafd08fe426357cc785c432a0b764