[mathml] Store mspace height in computed style
Previously we used vertical-align and SetInlineStyleProperty
to pass mspace height to the mspace layout algorithm, which was
a bit of a hack and due to SetInlineStyleProperty probably
caused security bug 1054765.
To fix this store mspace height as math baseline on computed style, setting it in the style adjuster.

Bug: 6606, 1054765
Change-Id: I87b7be8b5ed6d0b66b62ed9efb223cb4fc51f696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2068599
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Rob Buis <rbuis@igalia.com>
Cr-Original-Commit-Position: refs/heads/master@{#743903}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 038e787baec8a255485dc3310e765d77e384591e