[PerformanceObserver] Fix observe() edge cases
Currently, two calls to observe with the same 'type' will result in a
console error, whereas the spec implies that this is valid, with the
later call overriding the initial one. This CL fixes that. It also fixes
a problem where disconnect would not clear the filter_options_, as shown
in https://github.com/w3c/performance-timeline/pull/161. This would not
be a problem with 'entryTypes' because this variable would be replaced
on every call to observe(). Tests are added for these two cases.

Bug: 1052509
Change-Id: I75865b2925e2fb1a7e7fd64d7e979907c3b7650c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2067250
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#744012}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1303c01ebf362cb0dd006e41d1815c9cc24d35e5