Removing software usage fallback for CanvasResourceProvider::Create
This Cl removes the softwareUsage fallback for CanvasResourceProvider,
it adds a new set of constructors for the different type of Resource,
and replaces the calls for all instances of SoftwareUsage.

This is the first part of a big refactor removing all of the fallback
use cases and that will enable an overall simplification of the codebase
and it will make more explicit this part of the code base.


Bug: 1035589, 1021287
Change-Id: Iaeec2ee05d2a95f35eec753acc39e948260b3142
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2013483
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#744035}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1ee2162111a1d7098e7fd56a3d23d0f27d216708