[CSSRenderSubtree]: Globally force all activatable locks when doing find-in-page.
This patch forces all activatable locks in case of a find-in-page query. This
is a stop-gap measure to ensure sites that heavily rely on css-render-subtree
for performance can also be used with find-in-page reliably.

There is a longer term goal of having a better find-in-page which considers
the fact that some layout state might not be available (crbug.com/1045517)

However, since it is likely that render-subtree is going to be available
before the bug is fixed, we're landing this patch for now.

R=chrishtr@chromium.org, rakina@chromium.org

Change-Id: I67953e2f82910dd8ca52cb3b6d7d4ba99acd82d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2065052
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#744345}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 780b3f07bdce99d100979f07dfa69ab5188f91a2