CanvasRenderingContext2D should update font on demand
CanvasRenderingContext2D stores a font to draw text, but doesn't always
update it when the document's font cache is invalidated. In particular,
it's not notified when @font-face rules are added or removed.

This patch makes CanvasRenderingContext2D to ensure font validity when
using the font, by updating the font if it's invalid.

Bug: 1051715
Change-Id: I29d89a897b9ccaf14848dab0f33b115f4b3eee0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2053508
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#744363}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8eafe392d22e4e45c00ababce6c6de1782f2891b