Update wpt NFC tests to use mojo mocks for unavailable impl
This change updates the wpt tests in web-nfc/ to use the mockNFC object
in the tests that simulate a 'no implementation available' case.
MockNFC now has a new method simulateClosedPipe() which results in the
pipe being closed on a connection request.
This unblocks https://crrev.com/c/2062696, which adds code to crash the
renderer if no binder was found for a requested interface.

Bug: 1047680
Change-Id: Ib8f16f73a78982b2c17bf518d37d05e59ea384ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2072923
Commit-Queue: Oksana Zhuravlova <oksamyt@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#744760}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8907cc01b8d3947ccc7492d3f39c524752fdca1a