Add tracing event for long input delays
A trace is uploaded via "slow reports" when a long input delay
is detected (for opted-in clients). This trace event will make
it easier to identify the location of the long input delay when
analyzing the trace.

Bug: 1055890
Change-Id: Iba0e818b9c1a245da0099f44cc1224c437c8df97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2072665
Commit-Queue: Hailey Wang <haileywang@google.com>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#745064}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6cac30ce3c2c1192d19ba55f1eec519ea7010b29