Add GetSVGResourceAsType helpers
This adds helper functions for going from an SVGResource or
StyleSVGResource to the associated LayoutSVGResourceContainer type.
This reduces the amount of boilplate needed for this a little bit. These
types of access is expected to be little more common when the
SVGResources object has been phased out.

Add a special overload for ReferenceClipPathOperation that maps to a
LayoutSVGResourceClipper, and GetFilterResourceForSVG to capture the
current "single url(...)" restriction.

Bug: 1028063
Change-Id: I7d408a4d25d024162ace818a088b25be1205cedb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2072218
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#745090}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 321d124ad84da71ba5c24ae651a5d8f5a6379def