Use FCP instead of FMP in Time To Interactive
In this CL, we change the InteractiveDetector to rely on FCP instead of
FMP for TTI computations. In addition to this, the invalidating input
before FMP was only being used to determine whether to plumb TTI to the
browser process. As this is now never being done, that field can be
removed. The InteractiveDetectorTest is changed accordingly.

Bug: 1055806
Change-Id: Ie0154ffca8c79a0896679278c671e57668d0aa49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2076057
Reviewed-by: Steve Kobes <skobes@chromium.org>
Reviewed-by: Deep Roy <dproy@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#745137}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8cbfef6a969e6dacadf5b40f72e815a3307152fd