Menulist SELECT: Avoid updating UA shadow nodes in InsertedInto() and RemovedFrom()
HTMLOptionElement's InsertedInto() and RemovedFrom() called
MenuListSelectType::UpdateTextStyleAndContent(), which updates a Text
node in a UA shadow tree.  It triggered a DCHECK failure about
IsSlotAssignmentRecalcForbidden().  We should not update the DOM tree
in these functions.

This CL moves the logic of the InsertedInto() and the RemovedFrom() to
HTMLSelectElement::ChildrenChanged() and |HTMLOptGroupElement::
ChildrenChanged()|.

As for kAllChildrenRemoved such as |select.innerHTML=''|,
ChildrenChanged() had no ways to know removed nodes. This CL adds
ChildrenChangedAllChildrenRemovedNeedsList flag, and ContainerNode
provides a list of removed nodes if the flag is true.

This CL also updates comments on Node::InsertedInto(),
Node::RemovedFrom(), and ContainerNode::ChildrenChanged().

Bug: 1056094
Change-Id: I5a28a371b8bacd6551f17acfaf681032d615f1bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2074840
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#745230}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8581152742d5afb7b5101a639f43e4a630e96401