[video-raf] Add high fps auto-rescheduling
A lot of frames are missed when using video.rAF to paint high frame rate
videos (60fps). This is partially caused because, even if there is a new
frame ready to be painted, the calls to OnRequestAnimationFrame() after
jumping back on the main thread arrives too late, and we have to wait
until the next rendering steps.

This CL adds the logic to automatically re-schedule video.rAF execution
in the next rendering steps if we have a high enough frame rate, we have
more callbacks to run, and we consistently get new frames.

This CL might be revisited after collecting data from the canary builds.

Bug: 1012063
Change-Id: I90a63e32d774886c9e276b282ee105d1f359f968
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2079302
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#745747}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 892101572b2af3d40efcdf8dec711c1fec3e0776