Show CORS preflight request in DevTools
Currently CORS preflight OPTIONS requests are not shown in DevTools.
When OutOfBlinkCORS is not enabled, Blink’s ThreadableLoader fetches
the CORS preflight requests, and  the preflight request information is
sent to DevTools from ThreadableLoader
(http://trac.webkit.org/changeset/90340).
But when OutOfBlinkCORS is enabled, the CORS preflight requests are
fetched from the network process. And the preflight request
information is not sent to DevTools.

To fix this problem, this CL introduces 3 methods
(OnCorsPreflightRequest/Response/RequestCompleted) to
NetworkServiceClient mojo interface. PreflightLoader in the network
process calls these methods to pass the preflight request information
to the browser process. NetworkServiceClient in the browser process
receives the information and passes it to DevTools.

Bug: 941297
Change-Id: I56c1b5acd06ba06380986f054ea60473ab80ff52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2054468
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#745781}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3ca54b17528d9477f7738bf72970698337a7927c