[mathml] Implement API to access the OpenType MathVariants table
This CL implements an API to access data from [1] and adds very
basic unit tests for them. This might be refined later and will be used
to implement a shaper for stretchy operators [2].
OpenTypeMathStretchData is moved into its own small header so that it is
less intrusive when included in shape_result.h.

In order to simplify implementation, the number of parts and variants is
limited to a maximum of 20 which should be largely enough in practice.
Known open source fonts from [3] provide a maximum of PartCount=5 and
VariantCount=13 when executing the following command:

    rm -f out.txt
    for font in $(find -name '*.woff2'); do
      ttx -t MATH -o out.xml -f $font 2> /dev/null;
      egrep 'VariantCount|PartCount' out.xml >> out.txt ;
    done
    sort out.txt | uniq

[1] https://docs.microsoft.com/en-us/typography/opentype/spec/math#mathvariants-table
[2] https://chromium-review.googlesource.com/c/chromium/src/+/2051923
[3] https://fred-wang.github.io/MathFonts/

Bug: 6606
Change-Id: Ie1ce7bc0ae2790b1f6b48621b1cb806d395255dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2074678
Commit-Queue: Frédéric Wang <fwang@igalia.com>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#745895}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eb2efbb2880bc45d805a334ceeefc836531774f1