[Payment Handler] CanMakePaymentEvent.respondWithMinimalUI() #5
Patch #5: On desktop, pass currency to the CanMakePaymentEvent when
minimal UI feature is enabled, so the updated WPTs can pass.

Before this patch, the "currency" field of the CanMakePaymentEvent was
never set.

After this patch, the "currency" field of the CanMakePaymentEvent is set
on desktop when the minimal UI feature is enabled, e.g., through
chrome://flags/#enable-web-payments-minimal-ui flag, and the updated WPT
passes on desktop when opened in a full browser.

Bug: 1005076
Change-Id: Ib35ef8f808f57d7674a5712f12fbfe65918ce92a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2078797
Reviewed-by: Danyao Wang <danyao@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#745955}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a895796557a90b2d68b011c551986cc14daa59c0